backend/requirements.txt
Flask==3.0.2
jiwer==3.0.3
numpy==1.24.4
openai_whisper==20231117
pandas==2.0.3
pydub==0.25.1
python_Levenshtein==0.25.0
python_Levenshtein==0.25.0
torch==2.2.1
torchaudio==2.2.1
Werkzeug==3.0.1
gunicorn